Make pip-installable
