+++ .gitignore
@@ -1,0 +1,4 @@
+*~
+*.pyc
+*.egg-info
+__pycache__
