+++ environment.yaml
@@ -1,0 +1,211 @@
+name: FluViewer_v_0_1_x
+channels:
+- bioconda
+- conda-forge
+- pkgs/main
+dependencies:
+- _libgcc_mutex=0.1=conda_forge
+- _openmp_mutex=4.5=2_gnu
+- _sysroot_linux-64_curr_repodata_hack=3=h69a702a_13
+- alsa-lib=1.2.3.2=h166bdaf_0
+- bbmap=39.01=h92535d8_1
+- bcftools=1.17=h3cc50cf_1
+- bedtools=2.31.0=hf5e1c6e_2
+- blast=2.14.1=pl5321h6f7f691_0
+- brotli=1.0.9=h166bdaf_9
+- brotli-bin=1.0.9=h166bdaf_9
+- brotli-python=1.0.9=py38hfa26641_9
+- bwa=0.7.17=he4a0461_11
+- bzip2=1.0.8=h7f98852_4
+- c-ares=1.19.1=hd590300_0
+- ca-certificates=2023.7.22=hbcca054_0
+- cairo=1.16.0=h18b612c_1001
+- certifi=2023.7.22=pyhd8ed1ab_0
+- charset-normalizer=3.2.0=pyhd8ed1ab_0
+- clustalw=2.1=h4ac6f70_9
+- cmake=3.25.2=h077f3f9_0
+- contourpy=1.1.0=py38h7f3f72f_0
+- curl=7.87.0=h6312ad2_0
+- cycler=0.11.0=pyhd8ed1ab_0
+- eigen=3.2=3
+- entrez-direct=16.2=he881be0_1
+- expat=2.5.0=hcb278e6_1
+- fontconfig=2.14.2=h14ed4e7_0
+- fonttools=4.42.1=py38h01eb140_0
+- freebayes=1.3.6=hb0f3ef8_7
+- freetype=2.12.1=hca18f0e_1
+- gettext=0.21.1=h27087fc_0
+- giflib=5.2.1=h0b41bf4_3
+- glib=2.66.3=h58526e2_0
+- graphite2=1.3.13=h58526e2_1001
+- gsl=2.7=he838d99_0
+- harfbuzz=2.4.0=h37c48d4_1
+- htslib=1.17=h6bc39ce_1
+- icu=58.2=hf484d3e_1000
+- idna=3.4=pyhd8ed1ab_0
+- importlib-resources=6.0.1=pyhd8ed1ab_0
+- importlib_resources=6.0.1=pyhd8ed1ab_0
+- jpeg=9e=h0b41bf4_3
+- jsoncpp=1.9.5=h4bd325d_1
+- kernel-headers_linux-64=3.10.0=h4a8ded7_13
+- keyutils=1.6.1=h166bdaf_0
+- kiwisolver=1.4.5=py38h7f3f72f_0
+- krb5=1.20.1=hf9c8cef_0
+- lcms2=2.15=hfd0df8a_0
+- ld_impl_linux-64=2.40=h41732ed_0
+- lerc=4.0.0=h27087fc_0
+- libblas=3.9.0=17_linux64_openblas
+- libbrotlicommon=1.0.9=h166bdaf_9
+- libbrotlidec=1.0.9=h166bdaf_9
+- libbrotlienc=1.0.9=h166bdaf_9
+- libcblas=3.9.0=17_linux64_openblas
+- libcurl=7.87.0=h6312ad2_0
+- libdeflate=1.17=h0b41bf4_0
+- libedit=3.1.20191231=he28a2e2_2
+- libev=4.33=h516909a_1
+- libexpat=2.5.0=hcb278e6_1
+- libffi=3.2.1=he1b5a44_1007
+- libgcc-ng=13.1.0=he5830b7_0
+- libgfortran-ng=13.1.0=h69a702a_0
+- libgfortran5=13.1.0=h15d22d2_0
+- libglib=2.66.3=hbe7bbb4_0
+- libgomp=13.1.0=he5830b7_0
+- libiconv=1.17=h166bdaf_0
+- libidn2=2.3.4=h166bdaf_0
+- liblapack=3.9.0=17_linux64_openblas
+- libnghttp2=1.51.0=hdcd2b5c_0
+- libnsl=2.0.0=h7f98852_0
+- libopenblas=0.3.23=pthreads_h80387f5_0
+- libpng=1.6.39=h753d276_0
+- libsqlite=3.43.0=h2797004_0
+- libssh2=1.10.0=haa6b8db_3
+- libstdcxx-ng=13.1.0=hfd8a6a1_0
+- libtiff=4.5.0=h6adf6a1_2
+- libunistring=0.9.10=h7f98852_0
+- libuuid=2.38.1=h0b41bf4_0
+- libuv=1.44.2=hd590300_1
+- libwebp-base=1.3.1=hd590300_0
+- libxcb=1.15=h0b41bf4_0
+- libxml2=2.9.14=h74e7548_0
+- libzlib=1.2.13=hd590300_5
+- llvm-openmp=8.0.1=hc9558a2_0
+- matplotlib-base=3.7.2=py38hf5b0b65_0
+- munkres=1.1.4=pyh9f0ad1d_0
+- ncbi-vdb=3.0.6=hdbdd923_0
+- ncurses=6.4=hcb278e6_0
+- numpy=1.24.4=py38h59b608b_0
+- olefile=0.46=pyh9f0ad1d_1
+- openjdk=11.0.8=hacce0ff_0
+- openjpeg=2.5.0=hfec8fc6_2
+- openmp=8.0.1=0
+- openssl=1.1.1v=hd590300_0
+- ossuuid=1.6.2=hf484d3e_1000
+- packaging=23.1=pyhd8ed1ab_0
+- pandas=2.0.3=py38h01efb38_1
+- parallel=20160622=1
+- patsy=0.5.3=pyhd8ed1ab_0
+- pbzip2=1.1.13=0
+- pcre=8.45=h9c3ff4c_0
+- perl=5.32.1=4_hd590300_perl5
+- perl-alien-build=2.48=pl5321hec16e2b_0
+- perl-alien-libxml2=0.17=pl5321hec16e2b_0
+- perl-archive-tar=2.40=pl5321hdfd78af_0
+- perl-business-isbn=3.007=pl5321hdfd78af_0
+- perl-business-isbn-data=20210112.006=pl5321hdfd78af_0
+- perl-capture-tiny=0.48=pl5321hdfd78af_2
+- perl-carp=1.38=pl5321hdfd78af_4
+- perl-common-sense=3.75=pl5321hdfd78af_0
+- perl-compress-raw-bzip2=2.201=pl5321h87f3376_1
+- perl-compress-raw-zlib=2.105=pl5321h87f3376_0
+- perl-constant=1.33=pl5321hdfd78af_2
+- perl-data-dumper=2.183=pl5321hec16e2b_1
+- perl-encode=3.19=pl5321hec16e2b_1
+- perl-exporter=5.72=pl5321hdfd78af_2
+- perl-exporter-tiny=1.002002=pl5321hdfd78af_0
+- perl-extutils-makemaker=7.70=pl5321hd8ed1ab_0
+- perl-ffi-checklib=0.28=pl5321hdfd78af_0
+- perl-file-chdir=0.1010=pl5321hdfd78af_3
+- perl-file-path=2.18=pl5321hd8ed1ab_0
+- perl-file-temp=0.2304=pl5321hd8ed1ab_0
+- perl-file-which=1.24=pl5321hd8ed1ab_0
+- perl-importer=0.026=pl5321hdfd78af_0
+- perl-io-compress=2.201=pl5321hdbdd923_2
+- perl-io-zlib=1.14=pl5321hdfd78af_0
+- perl-json=4.10=pl5321hdfd78af_0
+- perl-json-xs=2.34=pl5321h4ac6f70_6
+- perl-list-moreutils=0.430=pl5321hdfd78af_0
+- perl-list-moreutils-xs=0.430=pl5321h031d066_2
+- perl-mime-base64=3.16=pl5321hec16e2b_2
+- perl-parent=0.236=pl5321hdfd78af_2
+- perl-path-tiny=0.122=pl5321hdfd78af_0
+- perl-pathtools=3.75=pl5321hec16e2b_3
+- perl-scalar-list-utils=1.62=pl5321hec16e2b_1
+- perl-scope-guard=0.21=pl5321hdfd78af_3
+- perl-sub-info=0.002=pl5321hdfd78af_1
+- perl-term-table=0.016=pl5321hdfd78af_0
+- perl-test2-suite=0.000145=pl5321hdfd78af_0
+- perl-threaded=5.32.1=hdfd78af_1
+- perl-types-serialiser=1.01=pl5321hdfd78af_0
+- perl-uri=5.12=pl5321hdfd78af_0
+- perl-xml-libxml=2.0207=pl5321h661654b_0
+- perl-xml-namespacesupport=1.12=pl5321hdfd78af_1
+- perl-xml-sax=1.02=pl5321hdfd78af_1
+- perl-xml-sax-base=1.09=pl5321hdfd78af_1
+- pillow=8.3.1=py38h2c7a002_0
+- pip=23.2.1=pyhd8ed1ab_0
+- pixman=0.38.0=h516909a_1003
+- platformdirs=3.10.0=pyhd8ed1ab_0
+- pooch=1.7.0=pyha770c72_3
+- pthread-stubs=0.4=h36c2ea0_1001
+- pyparsing=3.0.9=pyhd8ed1ab_0
+- pysocks=1.7.1=pyha2e5f31_6
+- python=3.8.5=h1103e12_9_cpython
+- python-dateutil=2.8.2=pyhd8ed1ab_0
+- python-tzdata=2023.3=pyhd8ed1ab_0
+- python_abi=3.8=3_cp38
+- pytz=2023.3=pyhd8ed1ab_0
+- readline=8.2=h8228510_1
+- requests=2.31.0=pyhd8ed1ab_0
+- rhash=1.4.3=hd590300_1
+- samtools=1.17=hd87286a_1
+- scipy=1.10.1=py38h59b608b_3
+- seaborn=0.12.2=hd8ed1ab_0
+- seaborn-base=0.12.2=pyhd8ed1ab_0
+- setuptools=68.1.2=pyhd8ed1ab_0
+- six=1.16.0=pyh6c4a22f_0
+- spades=3.15.3=h95f258a_1
+- sqlite=3.43.0=h2c6b66d_0
+- statsmodels=0.14.0=py38h31356c5_1
+- sysroot_linux-64=2.17=h4a8ded7_13
+- tabixpp=1.1.2=hd68fcf3_1
+- tk=8.6.12=h27826a3_0
+- typing-extensions=4.7.1=hd8ed1ab_0
+- typing_extensions=4.7.1=pyha770c72_0
+- unicodedata2=15.0.0=py38h0a891b7_0
+- urllib3=2.0.4=pyhd8ed1ab_0
+- vcflib=1.0.9=h146fbdb_2
+- wfa2-lib=2.3.3=h4ac6f70_1
+- wget=1.20.3=ha56f1ee_1
+- wheel=0.41.2=pyhd8ed1ab_0
+- xorg-fixesproto=5.0=h7f98852_1002
+- xorg-inputproto=2.3.2=h7f98852_1002
+- xorg-kbproto=1.0.7=h7f98852_1002
+- xorg-libice=1.1.1=hd590300_0
+- xorg-libsm=1.2.4=h7391055_0
+- xorg-libx11=1.8.6=h8ee46fc_0
+- xorg-libxau=1.0.11=hd590300_0
+- xorg-libxdmcp=1.1.3=h7f98852_0
+- xorg-libxext=1.3.4=h0b41bf4_2
+- xorg-libxfixes=5.0.3=h7f98852_1004
+- xorg-libxi=1.7.10=h7f98852_0
+- xorg-libxrender=0.9.11=hd590300_0
+- xorg-libxtst=1.2.3=h7f98852_1002
+- xorg-recordproto=1.14.2=h7f98852_1002
+- xorg-renderproto=0.11.1=h7f98852_1002
+- xorg-xextproto=7.3.0=h0b41bf4_1003
+- xorg-xproto=7.0.31=h7f98852_1007
+- xz=5.2.6=h166bdaf_0
+- zipp=3.16.2=pyhd8ed1ab_0
+- zlib=1.2.13=hd590300_5
+- zstd=1.5.5=hfc55251_0
+
